added particle shifting for dynamic window, added iterative buffer reduction, correction of the background buffer so it represents the air
the particle shifting makes the solution extremely stable.
The iterative buffer reduction is here to have better estimation of the density in the buffer. Without it an error is introduced due to the presence in the calculation of the density of the buffer particle that have a too high density ot be kept